Revert "Align snake and chess arena lighting and camera behavior"
