+++ webapp/src/components/SnakeBoard3D.jsx
@@ -1,6 +1,5 @@
 import { useEffect, useRef } from 'react';
 import * as THREE from 'three';
-import { ARENA_COLORS, createCarpetTextures } from './arenaMaterials.js';
 
 const clamp = (v, a, b) => Math.max(a, Math.min(b, v));
 
@@ -12,18 +11,25 @@ const WALL_PROXIMITY_FACTOR = 0.5;
 const WALL_HEIGHT_MULTIPLIER = 2;
 const CHAIR_SCALE = 4;
 const CHAIR_CLEARANCE = 0.52;
-const CAMERA_INITIAL_RADIUS_FACTOR = 1.35;
-const CAMERA_MIN_RADIUS_FACTOR = 0.95;
-const CAMERA_MAX_RADIUS_FACTOR = 2.4;
-const CAMERA_PHI_MIN = 0.92;
-const CAMERA_PHI_MAX = 1.22;
-const CAMERA_INITIAL_PHI_LERP = 0.35;
-const CAMERA_VERTICAL_SENSITIVITY = 0.003;
-const CAMERA_LEAN_STRENGTH = 0.0065;
+const CAMERA_INITIAL_RADIUS_FACTOR = 1.6;
+const CAMERA_MIN_RADIUS_FACTOR = 1.05;
+const CAMERA_MAX_RADIUS_FACTOR = 3.2;
+const CAMERA_ABS_MIN_PHI = 0.3;
+const STANDING_VIEW_PHI = 0.9;
+const CUE_SHOT_PHI = Math.PI / 2 - 0.26;
+const CAMERA_PHI_MIN = Math.max(CAMERA_ABS_MIN_PHI, STANDING_VIEW_PHI - 0.16);
+const CAMERA_PHI_MAX = CUE_SHOT_PHI - 0.24;
+const CAMERA_INITIAL_PHI_LERP = THREE.MathUtils.clamp(
+  (STANDING_VIEW_PHI - CAMERA_PHI_MIN) / (CAMERA_PHI_MAX - CAMERA_PHI_MIN || 1),
+  0,
+  1
+);
+const CAMERA_VERTICAL_SENSITIVITY = 0.0025;
+const CAMERA_LEAN_STRENGTH = 0.005;
 const CAMERA_CONFIG = {
-  fov: 52,
-  near: 0.1,
-  far: 5000
+  fov: 66,
+  near: 0.04,
+  far: 4000
 };
 
 const SNAKE_BOARD_TILES = 10;
@@ -45,6 +51,124 @@ const TOKEN_RADIUS = TILE_SIZE * 0.2;
 const TOKEN_HEIGHT = TILE_SIZE * 0.32;
 
 const DEFAULT_COLORS = ['#f97316', '#22d3ee', '#22c55e', '#a855f7'];
+
+const createCarpetTextures = (() => {
+  let cache = null;
+  const clamp01 = (v) => Math.min(1, Math.max(0, v));
+  const prng = (seed) => {
+    let value = seed;
+    return () => {
+      value = (value * 1664525 + 1013904223) % 4294967296;
+      return value / 4294967296;
+    };
+  };
+  const drawRoundedRect = (ctx, x, y, w, h, r) => {
+    const radius = Math.max(0, Math.min(r, Math.min(w, h) / 2));
+    ctx.beginPath();
+    ctx.moveTo(x + radius, y);
+    ctx.lineTo(x + w - radius, y);
+    ctx.quadraticCurveTo(x + w, y, x + w, y + radius);
+    ctx.lineTo(x + w, y + h - radius);
+    ctx.quadraticCurveTo(x + w, y + h, x + w - radius, y + h);
+    ctx.lineTo(x + radius, y + h);
+    ctx.quadraticCurveTo(x, y + h, x, y + h - radius);
+    ctx.lineTo(x, y + radius);
+    ctx.quadraticCurveTo(x, y, x + radius, y);
+    ctx.closePath();
+  };
+  return () => {
+    if (cache) return cache;
+    if (typeof document === 'undefined') {
+      cache = { map: null, bump: null };
+      return cache;
+    }
+
+    const size = 1024;
+    const canvas = document.createElement('canvas');
+    canvas.width = canvas.height = size;
+    const ctx = canvas.getContext('2d');
+
+    const gradient = ctx.createLinearGradient(0, 0, size, size);
+    gradient.addColorStop(0, '#7a0a18');
+    gradient.addColorStop(1, '#5e0913');
+    ctx.fillStyle = gradient;
+    ctx.fillRect(0, 0, size, size);
+
+    const rand = prng(987654321);
+    const image = ctx.getImageData(0, 0, size, size);
+    const data = image.data;
+    for (let y = 0; y < size; y++) {
+      for (let x = 0; x < size; x++) {
+        const idx = (y * size + x) * 4;
+        const fiber = (Math.sin((x / size) * Math.PI * 18) + Math.cos((y / size) * Math.PI * 22)) * 0.12;
+        const grain = (rand() - 0.5) * 0.22;
+        const shade = clamp01(0.96 + fiber + grain);
+        data[idx] = clamp01((data[idx] / 255) * shade) * 255;
+        data[idx + 1] = clamp01((data[idx + 1] / 255) * (0.98 + grain * 0.35)) * 255;
+        data[idx + 2] = clamp01((data[idx + 2] / 255) * (0.95 + grain * 0.2)) * 255;
+      }
+    }
+    ctx.putImageData(image, 0, 0);
+
+    ctx.globalAlpha = 0.05;
+    ctx.fillStyle = '#000000';
+    for (let row = 0; row < size; row += 3) {
+      ctx.fillRect(0, row, size, 1);
+    }
+    ctx.globalAlpha = 1;
+
+    const insetRatio = 0.055;
+    const stripeInset = size * insetRatio;
+    const stripeRadius = size * 0.08;
+    const stripeWidth = size * 0.012;
+    ctx.lineWidth = stripeWidth;
+    ctx.strokeStyle = '#d4af37';
+    ctx.shadowColor = 'rgba(0,0,0,0.18)';
+    ctx.shadowBlur = stripeWidth * 0.8;
+    drawRoundedRect(
+      ctx,
+      stripeInset,
+      stripeInset,
+      size - stripeInset * 2,
+      size - stripeInset * 2,
+      stripeRadius
+    );
+    ctx.stroke();
+    ctx.shadowBlur = 0;
+
+    const texture = new THREE.CanvasTexture(canvas);
+    texture.wrapS = texture.wrapT = THREE.ClampToEdgeWrapping;
+    texture.anisotropy = 8;
+    texture.minFilter = THREE.LinearMipMapLinearFilter;
+    texture.magFilter = THREE.LinearFilter;
+    texture.generateMipmaps = true;
+    if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+    else texture.encoding = THREE.sRGBEncoding;
+
+    const bumpCanvas = document.createElement('canvas');
+    bumpCanvas.width = bumpCanvas.height = size;
+    const bumpCtx = bumpCanvas.getContext('2d');
+    bumpCtx.drawImage(canvas, 0, 0);
+    const bumpImage = bumpCtx.getImageData(0, 0, size, size);
+    const bumpData = bumpImage.data;
+    for (let i = 0; i < bumpData.length; i += 4) {
+      const avg = (bumpData[i] + bumpData[i + 1] + bumpData[i + 2]) / 3;
+      const noise = (rand() - 0.5) * 32;
+      const value = clamp01((avg + noise) / 255) * 255;
+      bumpData[i] = bumpData[i + 1] = bumpData[i + 2] = value;
+    }
+    bumpCtx.putImageData(bumpImage, 0, 0);
+    const bump = new THREE.CanvasTexture(bumpCanvas);
+    bump.wrapS = bump.wrapT = THREE.ClampToEdgeWrapping;
+    bump.anisotropy = 4;
+    bump.minFilter = THREE.LinearMipMapLinearFilter;
+    bump.magFilter = THREE.LinearFilter;
+    bump.generateMipmaps = true;
+
+    cache = { map: texture, bump };
+    return cache;
+  };
+})();
 
 function createTileLabel(number) {
   const size = 128;
@@ -180,12 +304,8 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   const arena = new THREE.Group();
   scene.add(arena);
 
-  const ambient = new THREE.HemisphereLight(
-    ARENA_COLORS.hemisphereSky,
-    ARENA_COLORS.hemisphereGround,
-    ARENA_COLORS.hemisphereIntensity
-  );
-  arena.add(ambient);
+  const hemi = new THREE.HemisphereLight(0xffffff, 0x1a1f2b, 0.95);
+  arena.add(hemi);
   const key = new THREE.DirectionalLight(0xffffff, 1.0);
   key.position.set(1.8, 2.6, 1.6);
   arena.add(key);
@@ -207,7 +327,7 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   const floor = new THREE.Mesh(
     new THREE.PlaneGeometry(roomHalfWidth * 2, roomHalfDepth * 2),
     new THREE.MeshStandardMaterial({
-      color: ARENA_COLORS.floor,
+      color: 0x0f1222,
       roughness: 0.95,
       metalness: 0.05
     })
@@ -217,7 +337,7 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
 
   const carpetTextures = createCarpetTextures();
   const carpetMat = new THREE.MeshStandardMaterial({
-    color: ARENA_COLORS.carpet,
+    color: 0xb01224,
     roughness: 0.92,
     metalness: 0.04
   });
@@ -243,9 +363,9 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   const wallH = 3 * WALL_HEIGHT_MULTIPLIER;
   const wallT = 0.1;
   const wallMat = new THREE.MeshStandardMaterial({
-    color: ARENA_COLORS.wall,
-    roughness: ARENA_COLORS.wallRoughness,
-    metalness: ARENA_COLORS.wallMetalness,
+    color: 0xeeeeee,
+    roughness: 0.88,
+    metalness: 0.06,
     side: THREE.DoubleSide
   });
   const backWall = new THREE.Mesh(new THREE.BoxGeometry(halfRoomX * 2, wallH, wallT), wallMat);
@@ -264,7 +384,7 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   const ceilTrim = new THREE.Mesh(
     new THREE.BoxGeometry(halfRoomX * 2, 0.02, halfRoomZ * 2),
     new THREE.MeshStandardMaterial({
-      color: ARENA_COLORS.trim,
+      color: 0x1a233f,
       roughness: 0.9,
       metalness: 0.02,
       side: THREE.DoubleSide
@@ -274,8 +394,8 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   arena.add(ceilTrim);
 
   const ledMat = new THREE.MeshStandardMaterial({
-    color: ARENA_COLORS.led,
-    emissive: ARENA_COLORS.ledEmissive,
+    color: 0x00f7ff,
+    emissive: 0x0099aa,
     emissiveIntensity: 0.4,
     roughness: 0.6,
     metalness: 0.2,
@@ -365,104 +485,15 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
     return g;
   }
 
+  const chairA = makeChair();
   const seatHalfDepth = 0.25 * CHAIR_SCALE;
   const chairDistance = TABLE_TOP_SIZE / 2 + seatHalfDepth + CHAIR_CLEARANCE;
-
-  const chairsGroup = new THREE.Group();
-  arena.add(chairsGroup);
-
-  const chairs = Array.from({ length: 4 }, () => {
-    const chair = makeChair();
-    chair.visible = false;
-    chairsGroup.add(chair);
-    return chair;
-  });
-
-  const seatTokensGroup = new THREE.Group();
-  seatTokensGroup.position.set(
-    0,
-    tableSurfaceY + TABLE_TOP_THICKNESS / 2 + 0.02,
-    0
-  );
-  table.add(seatTokensGroup);
-
-  const makeSeatToken = () => {
-    const material = new THREE.MeshStandardMaterial({
-      color: new THREE.Color('#ffffff'),
-      roughness: 0.35,
-      metalness: 0.25
-    });
-    material.emissive.setHex(0x000000);
-    const group = new THREE.Group();
-    const base = new THREE.Mesh(
-      new THREE.CylinderGeometry(
-        TOKEN_RADIUS * 0.9,
-        TOKEN_RADIUS * 1.1,
-        TOKEN_HEIGHT * 0.45,
-        24
-      ),
-      material
-    );
-    base.position.y = TOKEN_HEIGHT * 0.225;
-    group.add(base);
-    const head = new THREE.Mesh(
-      new THREE.SphereGeometry(TOKEN_RADIUS * 0.65, 18, 14),
-      material
-    );
-    head.position.y = TOKEN_HEIGHT * 0.55;
-    group.add(head);
-    group.visible = false;
-    group.userData = { material };
-    return group;
-  };
-
-  const seatTokens = Array.from({ length: 4 }, () => {
-    const token = makeSeatToken();
-    seatTokensGroup.add(token);
-    return token;
-  });
-
-  const updateSeating = (playerList = []) => {
-    const activeCount = Math.max(2, Math.min(4, playerList.length || 0));
-    const baseAngle = -Math.PI / 2;
-    const angleStep = (Math.PI * 2) / activeCount;
-    for (let i = 0; i < chairs.length; i++) {
-      const chair = chairs[i];
-      if (i < activeCount) {
-        const angle = baseAngle + angleStep * i;
-        const x = Math.cos(angle) * chairDistance;
-        const z = Math.sin(angle) * chairDistance;
-        chair.visible = true;
-        chair.position.set(x, 0, z);
-        chair.rotation.y = Math.atan2(-x, -z);
-      } else {
-        chair.visible = false;
-      }
-
-      const seatToken = seatTokens[i];
-      const player = playerList[i];
-      if (player) {
-        seatToken.visible = true;
-        const tokenAngle = baseAngle + angleStep * i;
-        const tokenRadius = TABLE_TOP_SIZE / 2 - 0.22;
-        seatToken.position.set(
-          Math.cos(tokenAngle) * tokenRadius,
-          0,
-          Math.sin(tokenAngle) * tokenRadius
-        );
-        const mat = seatToken.userData.material;
-        const color = new THREE.Color(
-          player.color || DEFAULT_COLORS[i % DEFAULT_COLORS.length]
-        );
-        mat.color.copy(color);
-        mat.emissive.setHex(0x000000);
-      } else {
-        seatToken.visible = false;
-      }
-    }
-  };
-
-  updateSeating([]);
+  chairA.position.set(0, 0, -chairDistance);
+  arena.add(chairA);
+  const chairB = makeChair();
+  chairB.position.set(0, 0, chairDistance);
+  chairB.rotation.y = Math.PI;
+  arena.add(chairB);
 
   function makeStudioCamera() {
     const cam = new THREE.Group();
@@ -544,25 +575,15 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
   studioCamA.lookAt(boardLookTarget);
   studioCamB.lookAt(boardLookTarget);
 
-  const spotlight = new THREE.SpotLight(0xffffff, 0.65, 30, Math.PI / 4, 0.45, 1.2);
-  spotlight.position.set(0, boardLookTarget.y + 6, 0);
-  const spotTarget = new THREE.Object3D();
-  spotTarget.position.copy(boardLookTarget);
-  arena.add(spotlight);
-  arena.add(spotTarget);
-  spotlight.target = spotTarget;
-
   const camera = new THREE.PerspectiveCamera(
     CAMERA_CONFIG.fov,
     1,
     CAMERA_CONFIG.near,
     CAMERA_CONFIG.far
   );
-  const minR = SNAKE_BOARD_SIZE * CAMERA_MIN_RADIUS_FACTOR;
-  const maxR = SNAKE_BOARD_SIZE * CAMERA_MAX_RADIUS_FACTOR;
   const initialRadius = Math.max(
     SNAKE_BOARD_SIZE * CAMERA_INITIAL_RADIUS_FACTOR,
-    minR + 0.6
+    SNAKE_BOARD_SIZE * CAMERA_MIN_RADIUS_FACTOR
   );
   const sph = new THREE.Spherical(
     initialRadius,
@@ -578,6 +599,8 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
     camera.updateProjectionMatrix();
     const needed =
       SNAKE_BOARD_SIZE / (2 * Math.tan(THREE.MathUtils.degToRad(CAMERA_CONFIG.fov) / 2));
+    const minR = SNAKE_BOARD_SIZE * CAMERA_MIN_RADIUS_FACTOR;
+    const maxR = SNAKE_BOARD_SIZE * CAMERA_MAX_RADIUS_FACTOR;
     sph.radius = clamp(Math.max(needed, sph.radius), minR, maxR);
     const offset = new THREE.Vector3().setFromSpherical(sph);
     camera.position.copy(boardLookTarget).add(offset);
@@ -606,6 +629,8 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
     const phiDelta = -dy * CAMERA_VERTICAL_SENSITIVITY;
     sph.phi = clamp(sph.phi + phiDelta, CAMERA_PHI_MIN, CAMERA_PHI_MAX);
     const leanDelta = dy * CAMERA_LEAN_STRENGTH;
+    const minR = SNAKE_BOARD_SIZE * CAMERA_MIN_RADIUS_FACTOR;
+    const maxR = SNAKE_BOARD_SIZE * CAMERA_MAX_RADIUS_FACTOR;
     sph.radius = clamp(sph.radius - leanDelta, minR, maxR);
     fit();
   };
@@ -613,8 +638,9 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
     drag.on = false;
   };
   const onWheel = (e) => {
-    const r = sph.radius || initialRadius;
-    sph.radius = clamp(r + e.deltaY * 0.2, minR, maxR);
+    const minR = SNAKE_BOARD_SIZE * CAMERA_MIN_RADIUS_FACTOR;
+    const maxR = SNAKE_BOARD_SIZE * CAMERA_MAX_RADIUS_FACTOR;
+    sph.radius = clamp(sph.radius + e.deltaY * 0.0025, minR, maxR);
     fit();
   };
   dom.addEventListener('mousedown', onDown);
@@ -635,7 +661,7 @@ function buildArena(scene, renderer, host, cameraRef, disposeHandlers) {
     dom.removeEventListener('wheel', onWheel);
   });
 
-  return { arena, boardGroup, boardLookTarget, fit, updateSeating };
+  return { arena, boardGroup, boardLookTarget, fit };
 }
 
 function buildSnakeBoard(boardGroup, disposeHandlers = []) {
@@ -1096,7 +1122,6 @@ export default function SnakeBoard3D({
   const mountRef = useRef(null);
   const cameraRef = useRef(null);
   const boardRef = useRef(null);
-  const arenaRef = useRef(null);
   const fitRef = useRef(() => {});
   const disposeHandlers = useRef([]);
   const railTextureRef = useRef(null);
@@ -1121,7 +1146,6 @@ export default function SnakeBoard3D({
     handlers.length = 0;
 
     const arena = buildArena(scene, renderer, mount, cameraRef, handlers);
-    arenaRef.current = arena;
     const board = buildSnakeBoard(arena.boardGroup, handlers);
     boardRef.current = { ...board, boardLookTarget: arena.boardLookTarget };
 
@@ -1129,7 +1153,6 @@ export default function SnakeBoard3D({
     snakeTextureRef.current = makeSnakeTexture();
 
     fitRef.current = arena.fit;
-    arena.updateSeating(players);
 
     renderer.setAnimationLoop(() => {
       if (cameraRef.current) {
@@ -1149,7 +1172,6 @@ export default function SnakeBoard3D({
       }
       railTextureRef.current?.dispose?.();
       snakeTextureRef.current?.dispose?.();
-      arenaRef.current = null;
       renderer.dispose();
       scene.traverse((obj) => {
         if (obj.isMesh) {
@@ -1172,10 +1194,6 @@ export default function SnakeBoard3D({
       }
     }
   }, [highlight, trail, offsetPopup]);
-
-  useEffect(() => {
-    arenaRef.current?.updateSeating(players);
-  }, [players]);
 
   useEffect(() => {
     if (!boardRef.current) return;
